Fix negative matching (#14)
Fix #13 

The `shouldNegate` option was wrapping a negative pattern, e.g. `!pattern`, as a negative lookahead regular expression, e.g. `/^(?!pattern)$/`. When something that should pass was tested, like `unicorn`, the regular expression did not find `pattern`, but it also didn’t find anything that matched against `unicorn`, so the test failed when it shouldn’t.

This fix instead simply delegates the negation to `.isMatch()` (as was already the case for the default export) and inverts the return value if the `.negated` property returned from `makeRe()` is true.